+++ src/main/java/com/verdantartifice/primalmagick/common/enchantments/GuillotineEnchantment.java
@@ -1,0 +1,38 @@
+package com.verdantartifice.primalmagick.common.enchantments;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.AxeItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+/**
+ * Definition of an enchantment that gives some entities a chance to drop their heads when killed.
+ * 
+ * @author Daedalus4096
+ */
+public class GuillotineEnchantment extends AbstractRuneEnchantment {
+    public GuillotineEnchantment(Enchantment.Rarity rarity, EquipmentSlot... slots) {
+        super(rarity, EnchantmentCategory.WEAPON, slots);
+    }
+    
+    @Override
+    public int getMinCost(int enchantmentLevel) {
+        return 5 + ((enchantmentLevel - 1) * 10);
+    }
+    
+    @Override
+    public int getMaxCost(int enchantmentLevel) {
+        return this.getMinCost(enchantmentLevel) + 15;
+    }
+    
+    @Override
+    public int getMaxLevel() {
+        return 5;
+    }
+    
+    @Override
+    public boolean canEnchant(ItemStack stack) {
+        return stack.getItem() instanceof AxeItem ? true : super.canEnchant(stack);
+    }
+}